More pruning.
We're never going to use async io, and it's untested so it's out.
In the process of disabling relaxed ordering, I found that the
config registers were woefully bereft of content, so I filled that
in from the Xilinx documentation. Was also able to verify that the
disabling of relaxed ordering did what I expected (sadly, that did
not fix the bug I'm currently stuck on).